+++ docs/changes.rst
@@ -24,6 +24,7 @@ v1.1.6 (in development)
 - Update: ``basicinstructions`` has a new schedule.
 - Update: ``bgobt`` is no longer published.
 - Update: ``boasas`` is no longer published.
+- Update: ``bunny`` is no longer published.
 - Update: ``dieselsweetiesweb`` to be more robust to missing elements in the
   feed.